+++ packages/ui5-config/src/defaults.ts
@@ -9,6 +9,9 @@ export const enum UI5_DEFAULT {
     TYPES_VERSION_SINCE = '1.76.0',
     ESM_TYPES_VERSION_SINCE = '1.94.0',
     TYPES_VERSION_BEST = '1.108.0',
+    NEW_TYPES_PACKAGE_SINCE = '1.113.0',
     MANIFEST_VERSION = '1.12.0',
-    BASE_COMPONENT = 'sap/ui/core/UIComponent'
+    BASE_COMPONENT = 'sap/ui/core/UIComponent',
+    TS_TYPES_ESM_PACKAGE_NAME = '@sapui5/ts-types-esm',
+    TYPES_PACKAGE_NAME = '@sapui5/types'
 }