+++ packages/ui5-config/src/utils.ts
@@ -23,8 +23,8 @@ export function mergeObjects<B, E>(base: B, extension: E): B & E {
 
 /**
  * Get the best types version for the given minUI5Version within a selective range, starting at 1.90.0
- * for https://www.npmjs.com/package/@sapui5/ts-types-esm. For anything before use 1.90.0, and if no
- * minVersion is provided, use the latest LTS version (1.108.x).
+ * for https://www.npmjs.com/package/@sapui5/ts-types-esm or https://www.npmjs.com/package/@sapui5/types.
+ * For anything before use 1.90.0, and if no minVersion is provided, use the latest LTS version (1.108.x).
  *
  * @param minUI5Version the minimum UI5 version that needs to be supported
  * @returns semantic version representing the types version.
@@ -54,5 +54,22 @@ export function getTypesVersion(minUI5Version?: string) {
         return `~${UI5_DEFAULT.TYPES_VERSION_SINCE}`;
     } else {
         return `~${semVer.major(version)}.${semVer.minor(version)}.${semVer.patch(version)}`;
+    }
+}
+
+/**
+ * Get the correct type package name for the given ui5Version.
+ * For anything before 1.113.0, use https://www.npmjs.com/package/@sapui5/ts-types-esm,
+ * otherwise use the new package https://www.npmjs.com/package/@sapui5/types.
+ *
+ * @param ui5Version the UI5 version to get the correct package for
+ * @returns string representing the types package name.
+ */
+export function getTypesPackage(ui5Version?: string) {
+    const version = semVer.coerce(ui5Version) ?? semVer.coerce(UI5_DEFAULT.TYPES_VERSION_BEST);
+    if (version && semVer.lt(version, UI5_DEFAULT.NEW_TYPES_PACKAGE_SINCE)) {
+        return UI5_DEFAULT.TS_TYPES_ESM_PACKAGE_NAME;
+    } else {
+        return UI5_DEFAULT.TYPES_PACKAGE_NAME;
     }
 }