+++ packages/ui5-config/test/utils.test.ts
@@ -1,5 +1,5 @@
 import { UI5_DEFAULT } from '../src/defaults';
-import { getEsmTypesVersion, getTypesVersion, mergeObjects } from '../src/utils';
+import { getEsmTypesVersion, getTypesPackage, getTypesVersion, mergeObjects } from '../src/utils';
 
 describe('mergeObjects', () => {
     const base = {
@@ -44,7 +44,7 @@ describe('mergeObjects', () => {
         expect(merged.scripts?.first).toBe(extension.scripts?.first);
     });
 });
-describe('getEsmTypesVersion, getTypesVersion', () => {
+describe('getEsmTypesVersion, getTypesVersion, getTypesPackage', () => {
     const esmTypesVersionSince = `~${UI5_DEFAULT.ESM_TYPES_VERSION_SINCE}`;
     const typesVersionBest = `~${UI5_DEFAULT.TYPES_VERSION_BEST}`;
     const minU5Version = `~${UI5_DEFAULT.TYPES_VERSION_SINCE}`;
@@ -85,13 +85,44 @@ describe('getEsmTypesVersion, getTypesVersion', () => {
         ['1.102-snapshot', '~1.102.0'],
         ['1.109-snapshot', '~1.109.0']
     ];
+    const getTypesPackageTestData: [any, string][] = [
+        [UI5_DEFAULT.MIN_UI5_VERSION, UI5_DEFAULT.TS_TYPES_ESM_PACKAGE_NAME],
+        ['1', UI5_DEFAULT.TS_TYPES_ESM_PACKAGE_NAME],
+        ['1.70.0', UI5_DEFAULT.TS_TYPES_ESM_PACKAGE_NAME],
+        ['1.71.1', UI5_DEFAULT.TS_TYPES_ESM_PACKAGE_NAME],
+        ['metadata', UI5_DEFAULT.TS_TYPES_ESM_PACKAGE_NAME],
+        [undefined, UI5_DEFAULT.TS_TYPES_ESM_PACKAGE_NAME],
+        // Following versions dont exist
+        ['1.72.0', UI5_DEFAULT.TS_TYPES_ESM_PACKAGE_NAME],
+        ['1.73.0', UI5_DEFAULT.TS_TYPES_ESM_PACKAGE_NAME],
+        ['1.74.0', UI5_DEFAULT.TS_TYPES_ESM_PACKAGE_NAME],
+        ['1.75.0', UI5_DEFAULT.TS_TYPES_ESM_PACKAGE_NAME],
+        ['1.83.0', UI5_DEFAULT.TS_TYPES_ESM_PACKAGE_NAME],
+        [UI5_DEFAULT.TYPES_VERSION_BEST, UI5_DEFAULT.TS_TYPES_ESM_PACKAGE_NAME],
+        ['1.109.1', UI5_DEFAULT.TS_TYPES_ESM_PACKAGE_NAME],
+        ['1.80-snapshot', UI5_DEFAULT.TS_TYPES_ESM_PACKAGE_NAME],
+        ['1.102-snapshot', UI5_DEFAULT.TS_TYPES_ESM_PACKAGE_NAME],
+        ['1.109-snapshot', UI5_DEFAULT.TS_TYPES_ESM_PACKAGE_NAME],
+        // Following versions should use the new package
+        ['1.113.0', UI5_DEFAULT.TYPES_PACKAGE_NAME],
+        ['1.114.0', UI5_DEFAULT.TYPES_PACKAGE_NAME]
+    ];
+
     // Tests validation of versions against known versions https://www.npmjs.com/package/@sapui5/ts-types-esm
-    test.each(tesTSTypesEsmData)('Types version for @sapui5/ts-types-esm: (%p, %p)', (input, expected) => {
-        expect(getEsmTypesVersion(input)).toEqual(expected);
-    });
+    test.each(tesTSTypesEsmData)(
+        'Types version for @sapui5/ts-types-esm and @sapui5/types: (%p, %p)',
+        (input, expected) => {
+            expect(getEsmTypesVersion(input)).toEqual(expected);
+        }
+    );
 
     // Tests validation of versions against known versions https://www.npmjs.com/package/@sapui5/ts-types
     test.each(testSTypesData)('Types version for @sapui5/ts-types: (%p, %p)', (input, expected) => {
         expect(getTypesVersion(input)).toEqual(expected);
+    });
+
+    // Tests validation of getting the correct types package name
+    test.each(getTypesPackageTestData)('Types package: (%p, %p)', (input, expected) => {
+        expect(getTypesPackage(input)).toEqual(expected);
     });
 });