+++ examples/simple-generator/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @sap-ux/generator-simple-fe
 
+## 0.2.130
+
+### Patch Changes
+
+-   @sap-ux/fiori-elements-writer@0.12.20
+-   @sap-ux/fiori-freestyle-writer@0.19.13
+
+## 0.2.129
+
+### Patch Changes
+
+-   Updated dependencies [23059e62]
+    -   @sap-ux/axios-extension@1.3.6
+
 ## 0.2.128
 
 ### Patch Changes